requirements.txt
fastapi
librosa
tensorflow
transformers
torch
azure-cognitiveservices-speech
python-dotenv
threading